Refactor input circuit
